Implement Abstract Syntax Tree
The Abstract Syntax Tree (AST) provides an intermediate
representation of the syntax. The visitor pattern is used
to traverse the AST. Concrete visitor operations for each
node type are defined in Interpreter.rb

Again I've had to keep the file structure flat for the
meantime. I have found that attempting to group similar
classes into directories causes extra complexity.

This complexity also bleeds into the specs. I have also
removed a lot if not all mocking i.e. `instance_double`
and stubbing `allow(x).to receive(y) { ... }`. I think
I misinterpreted the whole ethos of TDD as 'make sure your
classes are tested in isolation'. Therefore, the mocking
and stubbing created a lot of complexity and coupling to
the code under test.

So it seems more correct to say 'make sure you can run
your specs in isolation', which can of course be done
without mocking and stubbing.

There may be a clearer approach to implementing the visitor
pattern also.