Compartmentalise Lexer logic
This will pull common lexer logic into its
own directory. This move also views the token
class has part of the lexer logic, which I think
makes sense.

I have also tried to simplify some of the parser
syntax/logic. For example

consume(TOKENS[:integer][:type])

is now refactored to the following

consume(plus_token)

This is achieved by using that 'metaprogramming'
craic. Hopefully I haven't over done it (?).

There are now explict references or 'requires' to
the classes that are required for the specs. This
was done to 'clean up' the top of the classes in
lib/.