Create new visitors and modularise executable
New visitor methods have been added to the interpreter
to be used to traverse the AST.

The main executable has been extended to accept a filename
from which would contain the source code of the new
language.

hello.plk now has actual source code in it that is runnable.
Finally the Makefile's 'run' target has been modified to
accept a file to run instead of running the REPL.